// src-backup-20250911-090210/lib/types.ts
export type PlayerData = {
    id: string;
    name: string;
    avatar: string;
    ltv: number;
    lastSeen: string;
    churnRisk: 'Низкий' | 'Средний' | 'Высокий';
    status: 'Активен' | 'Спящий' | 'Отток';
};